Bot code refactor + server announce fetch refactor
- Split code that was not directly related to the bot function (events, commands) to own files. Settings file containing settings, util file containing functions that are called on events.

- Refactored server announcement functions. update_server_announce function will update the announcements for a server using the server's key in the SERVER_ANNOUNCEMENT_CHANNEL_IDS dictionary as an argument/for identification. update_all_server_announce will simply run this function for each server.
- Moved check for whether a message is mentioning a server to the bot event.

- Changed update_announcements to update_global_announcements for clarification

NOTE!!! I have not tested this commit yet, more work will likely be done.
